Downgrade bundler until 1.5.2 is available
Bundler sets the THOR_DEBUG environment variable causing Thor to
re-raise exceptions on exit. This is breaking the railties tests
for generators where we are trying to capture error output.

The PR bundler/bundler#2794 fixes this by adding a per-instance
config `:debug` but this has not been released yet so to prevent
errors going unnoticed downgrade bundler to 1.3.5 until such a
time that 1.5.2 has been released.