This completes the work done for #14408 and #11432, which adds the --model-name option when generating scaffold controllers.
What was missing from the original work was correctly generating the route helpers to use when using the --model-name option.

This updates the path the generated code uses for redirection or links in the controller, views, and functional tests.

It also updates the generated views to render the form resource using a namespace, if it's necessary

It also adds better test coverage to the generated code.