Replace nil checks in quoting with a NullColumn.
This commit takes the nil check in the Connection adapters and replaces it with an
implementation of the NullObject pattern to avoid uncessary code branching. I would
have preferred to inject the NullColumn object directly, instead of using a nilguard,
but was unable to locate the code that passes a nil argument into the quote method.