Add ChronoModel to the DbConsole adapters list
ChronoModel http://github.com/ifad/chronomodel implements temporal extensions for ActiveRecord. As it decorates many migration methods, it defines a new adapter inheriting from the PostgreSQLAdapter.

However, as the Rails console command does not need to load any gem, there is no way for CM to hook into the dbconsole config, and adding it to the hardcoded list is necessary.