Add suffix to store_accessor.
An example where this is useful is the avatars store.
Having a column that stores the different sizes of an avatar like this
{ small: 'avatar1', medium: 'avatar2' } .
Having these methods exposed as user.small or user.medium does not make
any sense.
Adding the suffix '_avatar' brings clarity. user.small_avatar is
explicit enough.