String inflections should use I18n.locale if locale is not specified
When we set languages other than English (:en), we expect every word to
be translated to the language we set. However, String#pluralize and
String#singularize still use :en when locale is not specified. This
means we always need to specify locale every time we use one of these
methods.

To avoid that mess, use `I18n.locale` by default so everything will be
translated properly without specifying locale.