Consistency fix to the Getting Started with Rails guide
This corrects the code snippet taken from the app/views/posts/new.html.erb
file.  Earlier in the guide on lines 395-398 the user is instructed to change the
form_for line in the new.html.erb file to use the posts_path helper.  This fix 
updates the code snippet to be consistent with that earlier change.