Fix on email with inline attachments preview
If email is of type 'multipart/relative', then we render the `text/html`
part contained in it.

MailPreviews with inline attachments error out. They try to render an
HTML or plain text part, but it only sees the attachment part
(`image/png` for example) and a `multipart/relative`, returning an
unexpected array when views expect a `Message::Part`.

Fixes https://github.com/rails/rails/issues/14435