String to date conversions use current century
Previously, converting a String to Date with a 2 digit year would assume a year
between AD00 and AD99. As most software does not deal with these years, and it
is unlikely that this is considered expected behavior, `String#to_date` and
`String#to_datetime` now use the same defaults as `Date#parse`.