Adds support for Postgresql CIDR addresses, fixes #14857
The cidr format as specifies requries an unmasked address plus
a mask. Currently, we are reporting a masked address plus a mask.
This change correctly grabs the unmasked addr for a cidr and uses
the masked address for inet types.

This change depends on a change to the ruby lib/ipaddr.rb that is
currently here.

https://github.com/ruby/ruby/pull/599

Thanks to @maciekkolodziej for reporting.