Truncate day of month in TimeZone#parse.
As described in #14543, supplying missing components of a date from TimeZone#now
runs into problems when now.day is greater than the number of days in the month.
This can cause situations where parsing "February" when it's the 31st of January
causes TimeZone#parse to return a day in March.

This problem is resolved by returning the last day of the target month, thus
avoiding date-overflow issues.