Redirect links to api.rubyonrails.org
This changes the links in the read me for additonal information to point
to api.rubyonrails.org. It was previously using the link: option which
refer to local files whose path is relative to the --op directory
appending https://github.com/rails/rails/blob/master/activerecord/ to
the path and returning a 404 page.