Added 'all' environment to secrets.yml
Instead of making the user define a `key: value` in `secrets.yml` for
all
three environments, they can now set it in the `all` environment.
This `key: value` will be present in production, test, and development.
You
can override the key by defining it for a specific environment

For instance, with the `secrets.yml` file:

    development:
      my_key: my_val

    all:
      my_key: not_my_val

It would use `not_my_val` for test and production, and `my_val` in
development.

This is my first contribution -- if I've screwed up, just let me know!
Not sure if this change is something everyone wants, but I thought
it'd be nice for defining application critical API keys. I chose `all`
over
`global` because I felt it fit better, but it's a quick change if anyone
disagrees!