We need an explicit return
If we don't return early Ruby will memoize the value of the prefix of
the parent class what will make the subsequent searchs to not work as
expected.

If the early return we are avoiding the memoization.

But when using the deprecated path we need to memoize the value, so we
are not using early return for the deprecated path.