`TransactionState` is internal API, so added :nodoc:
This was introduced in 26853e8 / #9011. Its main purpose is to flip the
reference from `Transaction` -> AR objects to AR objects -> `TransactionState`.

This method this was extracted from originally from was a private API, and there
are no other public APIs to make this accessible to the user, so there is no
reason for this class to be public.

See also 67d8bb9 / #9068.