Added #upcase_first_char to upcase the first char leaving the rest as it is
The #capitalize method has the side effect of downcasing all the letters in a
string. This could not be the desired effect when we just want to upcase the
first letter.

This is particularly useful when the string contains acronyms or names and we
don't want to change the case of their letters.

e.g. "my name is Aldo and I like BBQ".capitalize returns
     "My name is aldo and i like bbq" while using #upcase_first_char becames
     "My name is Aldo and I like BBQ" which makes more sense.