Do not set integer columns to nil when assigning random values
This behavior was originally added to allow the following to work:

    Post.where(category_id: [1, 2]).new

This case will still continue to work as it did before, since this form
is likely used for things like access control. However, assigning types
which do not respond to `#to_i` to an integer column will now raise an
exception for the general case.