rm `serialized_attributes`
We have removed all code internally that relies on
`serialized_attributes`. The method exposes internals that shouldn't be
relied on, and all functionality it provides can be easily achieved
without this method. Remove the functionality, but leave a deprecation
warning on the method, in case there are gem providers who still need to
determine which columns are serialized.