Remove `type_cast_for_schema`
We already store the default value for a column in a database readable
format. If the value returned from that is good enough for the database
normally, it should be good enough for the database in this case as
well. With that in mind, the only special case remaining is the need for
quotes to be added to strings, which doesn't justify the method's
existence.