Added another guide to debugging Rails
I compiled about 32 pages of techniques and workflows for debugging Rails, intending for it to be the document I wish existed when I was an intermediate Rails programmer. I'd like to include this as a further reading link in the references section so that others can benefit from my write-up.