Fixes issue #15671
This allows both strings and symbols to be accepted as keys in a
nested association. Originally we attempted to fix this issue the
same as it was "fixed" on master, but that led to other issues of
backwards compatibility. We decided to fix this problem on 4-1-stable
by using symbols and not strings.

add tests for string and symbol keys for associations

These tests ensure that both strings and symbols are accepted
to retreive the association.