Fix redirect to properly handle urls of the form 'localhost:3000'
Absent this fix, 'localhost' is interpreted as the scheme and '3000' is interpreted as the 'opaque', in the sense of it being a directive to an unknown localhost scheme. Because Redirection is concerned with url redirection specifically, we can assume in these cases that the target is a url of a regular http / https scheme.