Add option to skip reseeding fixture data.
For those who use transactional fixtures, there's actually no real need
to truncate the test database and reseed the fixtures every time the
test suite runs, because the test data should only change when the
fixtures are updated.

This option allows them to turn off Rails automatic truncation and
reseeding of test data, speeding up the time it takes to run tests,
without losing any of the functionality of fixtures.