Include link_to show a post when index view created
I've worked through the Getting Started with Rails guide several times. Each time through I always get thrown when adding the link to show a specific post is added at the same time the link to edit a post is added to the index view. I updated the index view code in 5.8 to include adding the link to show a specific post.