Fixed regression with referencing polymorphic assoc in eager-load
This is cased by 03118bc + 9b5d603. The first commit referenced the undefined
local variable `column` when it should be using `reflection.type` as the lookup
key. The second commit changed `build_arel` to not modify the `bind_values` in-
place so we need to combine the arel's `bind_values` with the relation's when
building the SQL.

Fixes #16591

Related #15821 / #15892 / 7aeca50