Revert "inject Rack::Lock if config.eager_load is false"
After some discussion in Campfire we have decided to revert this one.
There are legit use cases where concurrent but coordinated requests
are made.

We have seen a test suite that uses Rack::Reproxy to redispatch
requests down, and also some actions that call un-mocked services
that call the original app and need to complete the cycle to
complete the original request.

While this is a more sane default for 4.2, its backport is more
dubious because it would break those test suites.

As a workaround, people having issues related to this before 4.2 can
throw

    config.allow_concurrency = false

in config/environments/test.rb.

This reverts commit 65f75c0ceadac0080a576c1a4551b7c55f4bdff6.